+++ Dockerfile
@@ -7,4 +7,4 @@ RUN apt-get update && apt install git unzip -y
 ENV COMPOSER_ALLOW_SUPERUSER 1
 RUN composer install 
 WORKDIR /usr/src/php-demo-serialize/src
-CMD [ "php", "MessageDTOTest.php" ]
+CMD [ "php", "MessageEntityTest.php" ]
