+++ Readme.md
@@ -22,6 +22,23 @@
 - UnitTest класс MessageEntityTest с двумя тест методами (пустой и с данными), критерий приемки $this->assertEquals($message, Message::deserialize($message->serialize());
 
 ## Реализация
-Каждый класс наследуется от класса, который предоставляет возможность сериализовать данные, которые хранятся в поле $objectData.
+Функционал по сохранению и восcтановлению свойств объекта (в том числе и мета-данных) обеспечивается встроенными функциями serialize и unserialize. Данные функции помещены в функции-обертки в абстрактном классе JsonDeserializator.
+
+Класс сообщений наследуется от JsonDeserializator и получает возможность сохранять и восстанавливать состояние.
+
+## Проверка
+Для проверки созданы две тестовые функции. Внутри функции с данными тестируется два сценария - на основе данных инициализации данных в конструкторе Message и при добавлении дополнительного поля, сохраняющего класс.
+После восстановления объекта так же проверяется работа метода из nested-класса (dynamic property).
 
 ## Использование
+Для проверки работоспособности предлагается использовать docker-контейнер.
+Сначала надо собрать контейнер:
+> docker build --pull --rm -f "Dockerfile" -t phpdemoserialize:latest "."
+
+Далее запустить контейнер и проверить результат исполнения теста:
+> docker run --rm -it  phpdemoserialize:latest 
+
+Ожидаемый вывод в консоли:
+> Testing is done
+
+Что говорит о том, что все тесты отработали корректно.
